commit_4

+++ client/src/comps/Navbar.js
@@ -1,19 +1,17 @@
-import React, {useContext, useEffect, useState} from 'react'
+import React, {useContext} from 'react'
 import {Link, NavLink} from "react-router-dom";
 import {FontAwesomeIcon} from '@fortawesome/react-fontawesome'
 import {faGlassCheers} from '@fortawesome/free-solid-svg-icons'
 import {Context} from "../index";
 import {useAuthState} from "react-firebase-hooks/auth";
+import {useHistory} from "react-router";
 
 export default function Navbar() {
     const {auth} = useContext(Context)
     const [user] = useAuthState(auth)
-    const [userData, setUserData] = useState(JSON.parse(localStorage.getItem('profile')))
+    const history = useHistory()
+    const userData = JSON.parse(localStorage.getItem('profile'))
     console.log(userData);
-
-    // useEffect(() => {
-    //     setUserData(userData)
-    // },[userData])
 
     return (
         <div>
@@ -24,7 +22,6 @@ export default function Navbar() {
                         <div className='icon ml-10'><FontAwesomeIcon icon={faGlassCheers}/></div>
                     </div>
                 </Link>
-
                 <ul className='d-flex mt-22'>
                     <li><NavLink to={'/pubs'} className='mr-20' activeClassName='isActive'>Pubs</NavLink></li>
                     <li><NavLink to={'/news'} className='mr-20' activeClassName='isActive'>News</NavLink></li>
@@ -42,8 +39,11 @@ export default function Navbar() {
                                     to={"/login"}
                                     className='ml-20 mr-20 isActive'
                                     onClick={() => {
-                                        auth.signOut();
-                                        localStorage.clear()
+                                        if (user) {
+                                            auth.signOut();
+                                        } else {
+                                            localStorage.clear()
+                                        }
                                     }}
                                 >Logout
                                 </NavLink>
